When converting to a string, always show Quantities to 2 decimal places.
I haven't been able to track down exactly what change in Python 3 causes this,
but when a Quantity with the value '1' is rendered in Python 2.7 the result is
'1' and in Python 3 the result is '1.0'.

Since there is no difference in correctness, this commit updates changes the
semantics of Quantity's str method so it always returns 2 decimal places. I
updated the tests accordingly, and now the test suites works in Python 2.5, 2.6,
2.7 and 3.3.